pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mirrorbase"
version = "0.1.0"
description = "Instant live clones of Postgres databases. Paste a URL, get a writable fork in seconds."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Teddy Oweh" },
]
keywords = ["postgres", "database", "clone", "copy-on-write", "replication"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
]
dependencies = [
    "click>=8.0",
    "psycopg2-binary>=2.9",
    "rich>=13.0",
]

[project.scripts]
mirrorbase = "mirrorbase.cli:cli"

[project.urls]
Homepage = "https://github.com/teddyoweh/mirrorbase"
Repository = "https://github.com/teddyoweh/mirrorbase"
Issues = "https://github.com/teddyoweh/mirrorbase/issues"
